apps/api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
pydantic>=2.0.0
cryptography>=43.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
redis>=5.0.0
